+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/Updated.java
@@ -1,0 +1,9 @@
+package io.github.ulisse1996.jaorm.dsl.query.common;
+
+import io.github.ulisse1996.jaorm.dsl.query.common.intermediate.IntermediateUpdate;
+import io.github.ulisse1996.jaorm.entity.SqlColumn;
+
+public interface Updated<T> {
+
+    <R> IntermediateUpdate<T, R> setting(SqlColumn<T, R> column);
+}