+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/UpdatedWhere.java
@@ -1,0 +1,6 @@
+package io.github.ulisse1996.jaorm.dsl.query.common;
+
+import io.github.ulisse1996.jaorm.dsl.query.common.trait.WithExecute;
+import io.github.ulisse1996.jaorm.dsl.query.common.trait.WithUpdatedWhere;
+
+public interface UpdatedWhere<T> extends WithUpdatedWhere<T>, WithExecute {}