+++ jaorm-dsl/src/main/java/io/github/ulisse1996/jaorm/dsl/query/common/trait/WithExecute.java
@@ -1,0 +1,6 @@
+package io.github.ulisse1996.jaorm.dsl.query.common.trait;
+
+public interface WithExecute {
+
+    void execute();
+}